requirements.txt
asgiref==3.5.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.11 
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713 
Django==4.0.2
django-cors-headers==3.11.0
django-excel==0.0.10       
django-filter==21.1        
django-import-export==2.7.1
django-rest-swagger==2.2.0 
djangorestframework==3.13.1
et-xmlfile==1.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
lml==0.1.0
Markdown==3.3.6
MarkupPy==1.14
MarkupSafe==2.0.1
mysqlclient==2.1.0
numpy==1.22.2
odfpy==1.4.1
openapi-codec==1.3.2       
openpyxl==3.0.9
pandas==1.4.0
pycodestyle==2.8.0
pyexcel==0.6.7
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyexcel-xlsx==0.6.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tablib==3.2.0
texttable==1.6.4
toml==0.10.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
xlrd==2.0.1
xlwt==1.3.0